after student review. Fruit class was created instead of Enums

+++ src/main/java/core/basesyntax/strategy/operation/OperationHandler.java
@@ -1,7 +1,5 @@
 package core.basesyntax.strategy.operation;
 
-import java.math.BigDecimal;
-
 public interface OperationHandler {
-    BigDecimal getShopOperation(BigDecimal amount);
+    Integer getShopOperation(Integer amount);
 }
