+++ src/main/java/core/basesyntax/strategy/operation/SupplyOperationHandler.java
@@ -1,10 +1,8 @@
 package core.basesyntax.strategy.operation;
 
-import java.math.BigDecimal;
-
 public class SupplyOperationHandler implements OperationHandler {
     @Override
-    public BigDecimal getShopOperation(BigDecimal amount) {
+    public Integer getShopOperation(Integer amount) {
         return amount;
     }
 }
